requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-RESTx==1.1.0
SQLAlchemy==2.0.23
Werkzeug==2.3.7
email-validator==2.1.0
WTForms==3.1.1
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
google-generativeai==0.8.3